Move locale files dir
