+++ i18n/en/htmlCharCodes.json
