+++ i18n/en/offline.json
@@ -1,0 +1,4 @@
+{
+  "heading": "Oof! Page not found!",
+  "message": "Choose a tool from the menu to get started!"
+}
