+++ i18n/ja/aspectRatio.json
