+++ i18n/ja/htmlCharCodes.json
