+++ i18n/ja/regex.json
